requirements.txt
flask>=2.3
flask-cors>=4.0
numpy>=1.24
torch>=2.1
mne>=1.6
PyWavelets>=1.4
matplotlib>=3.7
gunicorn>=21.2
PyYAML>=6.0
psutil>=5.9